# swagger/portal-api.json
{
    "swagger": "2.0",
    "info": {
        "version": "1.0.0",
        "title": "Västtrafik REST API for WSO2 API Manager",
        "description": "This document describe a ** RESTFul API ** for WSO2 **API Manager**.\n\n\nYou can find the source of this API definition at [here](https://github.com/vasttrafik/wso2-apim-developer-portal-api). It was written with [swagger 2](http://swagger.io/).\n",
        "contact": {
            "name": "Västtrafik",
            "url": "https://developer.vasttrafik.se",
            "email": "api@vasttrafik.se"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        }
    },
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ],
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/statistics": {
            "get": {
                "parameters": [
                    {
                        "in": "query",
                        "name": "source",
                        "description": "Source of statistics info. Either for API gateway or applications.",
                        "required": true,
                        "type": "string",
                        "enum": [
                            "apis",
                            "applications"
                        ]
                    },
                    {
                        "in": "query",
                        "name": "grouping",
                        "description": "Date grouping for statistics. Defaults to day.",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "day",
                            "week"
                        ]
                    },
                    {
                        "in": "query",
                        "name": "period",
                        "description": "Period for statistics. Defaults to week.",
                        "required": false,
                        "type": "string",
                        "enum": [
                            "week",
                            "month",
                            "quarter"
                        ]
                    }
                ],
                "description": "Retrieves combined statistics for chosen source",
                "responses": {
                    "200": {
                        "description": "OK statistics returned.",
                        "schema": {
                            "title": "StatisticsList",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Statistic"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Either the request did not contain a valid Authorization header, or the access token expired and needs to be refreshed.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/statistics/{applicationName}": {
            "parameters": [
                {
                    "name": "applicationName",
                    "in": "path",
                    "description": "The application name",
                    "required": true,
                    "type": "string"
                },
                {
                    "in": "query",
                    "name": "type",
                    "description": "Statistics type",
                    "required": true,
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "totalRequestsFaults",
                            "faultsPercentage",
                            "uniqueUsers",
                            "requestResponseTimes"
                        ],
                        "collectionFormat": "csv"
                    }
                }
            ],
            "get": {
                "description": "Retrieves statistics available for specified applicationId",
                "responses": {
                    "200": {
                        "description": "OK statistics returned.",
                        "schema": {
                            "title": "StatisticsList",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Statistic"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Either the request did not contain a valid Authorization header, or the access token expired and needs to be refreshed.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
        "/statistics/{apiName}/{apiVersion}": {
            "parameters": [
                {
                    "name": "apiName",
                    "in": "path",
                    "description": "The api name",
                    "required": true,
                    "type": "string"
                },
                {
                    "name": "apiVersion",
                    "in": "path",
                    "description": "The api version",
                    "required": true,
                    "type": "string"
                },
                {
                    "in": "query",
                    "name": "grouping",
                    "description": "Date grouping for statistics. Defaults to / day.",
                    "required": false,
                    "type": "string",
                    "enum": [
                        "day",
                        "week"
                    ]
                },
                {
                    "in": "query",
                    "name": "period",
                    "description": "Period for statistics. Defaults to week.",
                    "required": false,
                    "type": "string",
                    "enum": [
                        "week",
                        "month",
                        "quarter"
                    ]
                },
                {
                    "in": "query",
                    "name": "type",
                    "description": "Statistics type",
                    "required": true,
                    "type": "array",
                    "items": {
                        "type": "string",
                        "enum": [
                            "totalRequests",
                            "requestResponseTimes"
                        ],
                        "collectionFormat": "csv"
                    }
                }
            ],
            "get": {
                "description": "Retrieves statistics available for specified applicationId",
                "responses": {
                    "200": {
                        "description": "OK statistics returned.",
                        "schema": {
                            "title": "StatisticsList",
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/Statistic"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized. Either the request did not contain a valid Authorization header, or the access token expired and needs to be refreshed.",
                        "schema": {
                            "$ref": "#/definitions/Error"
                        }
                    }
                }
            }
        },
...
    },
...
    "definitions": {
...
        "Statistic": {
            "title": "Statistic",
            "properties": {
                "type": {
                    "type": "string"
                },
                "api": {
                    "type": "string"
                },
                "apiVersion": {
                    "type": "string"
                },
                "series": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Series"
                    }
                }
            }
        },
        "Series": {
            "title": "Series",
            "properties": {
                "name": {
                    "type": "string"
                },
                "values": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "additionalProperties": {
                            "type": "number"
                        }
                    }
                }
            }
        },
        "Error": {
            "title": "Error object returned with 4XX HTTP status",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int64"
                },
                "message": {
                    "type": "string",
                    "description": "Error message."
                },
                "description": {
                    "type": "string",
                    "description": "A detail description about the error message."
                },
                "moreInfo": {
                    "type": "string",
                    "description": "Preferably an url with more details about the error."
                },
                "items": {
                    "type": "array",
                    "description": "If there are more than one error list them out. Ex. list out validation errors by each field.",
                    "items": {
                        "$ref": "#/definitions/ErrorListItem"
                    }
                }
            }
        },
        "ErrorListItem": {
            "title": "Description of individual errors that may have occored during a request.",
            "required": [
                "code",
                "message"
            ],
            "properties": {
                "code": {
                    "type": "integer",
                    "format": "int64"
                },
                "message": {
                    "type": "string",
                    "description": "Description about individual errors occored"
                }
            }
        }
    }
}